build: Enable maximum compiler warnings with -Werror, fix all warnings
CMakeLists.txt now enforces 19 warning flags including:
  -Wall -Wextra -Wpedantic -Werror
  -Wshadow -Wconversion -Wsign-conversion
  -Wold-style-cast -Wcast-align -Wnull-dereference
  -Wdouble-promotion -Wformat=2 -Wimplicit-fallthrough
  + GCC-specific: -Wduplicated-cond -Wduplicated-branches -Wlogical-op

Fixed all warnings:
- import_builder: check/cast [[nodiscard]] return values
- layout: mark unused parameter, fix size_t→uint32_t narrowing
- coff_editor: fix sign/size conversion in symbol loading
- editor_test: fix narrowing casts and unchecked returns
- dump_symbols example: fix sign conversion

Zero warnings, zero errors, 91 tests passing.

+++ coffi-modern/include/coffi/editor/coff_editor.hpp
@@ -271,7 +271,7 @@ public:
                 lr->total_size = lr->symbol_table_off;
                 uint32_t sym_slots = 0;
                 for (auto& s : symbols_) sym_slots += s.total_slots();
-                lr->total_size += sym_slots * sizeof(symbol_record);
+                lr->total_size += sym_slots * static_cast<uint32_t>(sizeof(symbol_record));
                 if (!strings_.empty()) lr->total_size += strings_.size();
             }
         }
@@ -510,7 +510,7 @@ private:
                     if (aux) se.add_aux(*aux);
                 }
 
-                i += 1 + rec->aux_symbols_number;
+                i += 1u + rec->aux_symbols_number;
                 symbols_.push_back(std::move(se));
             }
